[LNA] Add error messages on chrome://policy for IP overrides
Add an PolicyHandler for LocalNetworkAccessIpAddressSpaceOverrides to
show error messages on the chrome://policy page for invalid IP address
space overrides.

Screenshot: https://issues.chromium.org/507804958#attachment76411206

changes to
components/policy/core/browser/configuration_policy_handler.cc

are because without the change the error messages looked like:

Error at LocalNetworkAccessIpAddressSpaceOverrides[0][0][0][0]: Value
doesn't match format.

because the type of PolicyErrorPath is

using PolicyErrorPath = std::vector<std::variant<int, std::string>>;

and doing the following:

size_t i = 4;
std::vector<std::variant<int, std::string>> v{i}

creates a vector of size 4

Bug: 507804958
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7823083
Commit-Queue: Hubert Chao <hchao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1630361}

+++ components/policy/core/browser/configuration_policy_handler.cc
@@ -153,17 +153,27 @@ bool ListPolicyHandler::CheckAndGetList(
     const base::Value& entry = list[list_index];
     if (entry.type() != list_entry_type_) {
       if (errors) {
-        errors->AddError(policy_name(), IDS_POLICY_TYPE_ERROR,
-                         base::Value::GetTypeName(list_entry_type_),
-                         PolicyErrorPath{list_index});
+        // Casting to int avoids a narrowing conversion from size_t when
+        // constructing PolicyErrorPath, which would otherwise cause the
+        // compiler to prefer the std::vector(size_t count) constructor over the
+        // initializer list constructor.
+        errors->AddError(
+            policy_name(), IDS_POLICY_TYPE_ERROR,
+            base::Value::GetTypeName(list_entry_type_),
+            PolicyErrorPath{base::saturated_cast<int>(list_index)});
       }
       continue;
     }
 
     if (!CheckListEntry(entry)) {
       if (errors) {
-        errors->AddError(policy_name(), IDS_POLICY_VALUE_FORMAT_ERROR,
-                         PolicyErrorPath{list_index});
+        // Casting to int avoids a narrowing conversion from size_t when
+        // constructing PolicyErrorPath, which would otherwise cause the
+        // compiler to prefer the std::vector(size_t count) constructor over the
+        // initializer list constructor.
+        errors->AddError(
+            policy_name(), IDS_POLICY_VALUE_FORMAT_ERROR,
+            PolicyErrorPath{base::saturated_cast<int>(list_index)});
       }
       continue;
     }
@@ -656,9 +666,13 @@ bool SimpleJsonStringSchemaValidatingPolicyHandler::CheckListOfJsonStrings(
     const base::Value& entry = list[index];
     if (!entry.is_string()) {
       if (errors) {
+        // Casting to int avoids a narrowing conversion from size_t when
+        // constructing PolicyErrorPath, which would otherwise cause the
+        // compiler to prefer the std::vector(size_t count) constructor over the
+        // initializer list constructor.
         errors->AddError(policy_name(), IDS_POLICY_TYPE_ERROR,
                          base::Value::GetTypeName(base::Value::Type::STRING),
-                         PolicyErrorPath{index});
+                         PolicyErrorPath{base::saturated_cast<int>(index)});
       }
       continue;
     }
